Moved menu to presentation and turned into a flexible implementation
